Fix issues after merge and split/clean up protocol in sub-files
