+++ pkg/protocol/engine/attestation/slotattestation/manager.go
@@ -104,9 +104,29 @@ func (m *Manager) Shutdown() {
 	m.TriggerStopped()
 }
 
-// Get returns the attestations that are included in the commitment of the given slot.
+// Get returns the attestations that are included in the commitment of the given slot as list.
 // If attestationOffset=3 and commitment is 10, then the returned attestations are blocks from 7 to 10 that commit to at least 7.
-func (m *Manager) Get(index iotago.SlotIndex) (*ads.Map[iotago.AccountID, iotago.Attestation, *iotago.AccountID, *iotago.Attestation], error) {
+func (m *Manager) Get(index iotago.SlotIndex) (attestations []*iotago.Attestation, err error) {
+	adsMap, err := m.GetMap(index)
+	if err != nil {
+		return nil, errors.Wrapf(err, "failed to get attestations for slot %d", index)
+	}
+
+	attestations = make([]*iotago.Attestation, 0)
+	if err := adsMap.Stream(func(_ iotago.AccountID, attestation *iotago.Attestation) bool {
+		attestations = append(attestations, attestation)
+
+		return true
+	}); err != nil {
+		return nil, errors.Wrapf(err, "failed to stream attestations for slot %d", index)
+	}
+
+	return attestations, nil
+}
+
+// GetMap returns the attestations that are included in the commitment of the given slot as ads.Map.
+// If attestationOffset=3 and commitment is 10, then the returned attestations are blocks from 7 to 10 that commit to at least 7.
+func (m *Manager) GetMap(index iotago.SlotIndex) (*ads.Map[iotago.AccountID, iotago.Attestation, *iotago.AccountID, *iotago.Attestation], error) {
 	m.commitmentMutex.RLock()
 	defer m.commitmentMutex.RUnlock()
