+++ pkg/protocol/protocol.go
@@ -6,8 +6,6 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/iotaledger/hive.go/ds/types"
-	"github.com/iotaledger/hive.go/lo"
 	"github.com/iotaledger/hive.go/runtime/event"
 	"github.com/iotaledger/hive.go/runtime/module"
 	"github.com/iotaledger/hive.go/runtime/options"
@@ -21,7 +19,6 @@ import (
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/attestation/slotattestation"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/blockdag"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/blockdag/inmemoryblockdag"
-	"github.com/iotaledger/iota-core/pkg/protocol/engine/blocks"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/booker"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/booker/inmemorybooker"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/clock"
@@ -176,55 +173,6 @@ func (p *Protocol) Shutdown() {
 	p.SyncManager.Shutdown()
 }
 
-func (p *Protocol) runNetworkProtocol() {
-	p.networkProtocol = core.NewProtocol(p.dispatcher, p.Workers.CreatePool("NetworkProtocol"), p.API()) // Use max amount of workers for networking
-	p.Events.Network.LinkTo(p.networkProtocol.Events)
-
-	wpBlocks := p.Workers.CreatePool("NetworkEvents.Blocks") // Use max amount of workers for sending, receiving and requesting blocks
-
-	p.Events.Network.BlockReceived.Hook(func(block *model.Block, id network.PeerID) {
-		if err := p.ProcessBlock(block, id); err != nil {
-			p.ErrorHandler()(err)
-		}
-	}, event.WithWorkerPool(wpBlocks))
-
-	p.Events.Network.BlockRequestReceived.Hook(func(blockID iotago.BlockID, id network.PeerID) {
-		if block, exists := p.MainEngineInstance().Block(blockID); exists {
-			p.networkProtocol.SendBlock(block, id)
-		}
-	}, event.WithWorkerPool(wpBlocks))
-
-	p.Events.Engine.BlockRequester.Tick.Hook(func(blockID iotago.BlockID) {
-		p.networkProtocol.RequestBlock(blockID)
-	}, event.WithWorkerPool(wpBlocks))
-
-	p.Events.Engine.BlockDAG.BlockSolid.Hook(func(block *blocks.Block) {
-		p.networkProtocol.SendBlock(block.ModelBlock())
-	}, event.WithWorkerPool(wpBlocks))
-
-	wpCommitments := p.Workers.CreatePool("NetworkEvents.SlotCommitments")
-
-	p.Events.Network.SlotCommitmentRequestReceived.Hook(func(commitmentID iotago.CommitmentID, source network.PeerID) {
-		// when we receive a commitment request, do not look it up in the ChainManager but in the storage, else we might answer with commitments we did not issue ourselves and for which we cannot provide attestations
-		if requestedCommitment, err := p.MainEngineInstance().Storage.Commitments().Load(commitmentID.Index()); err == nil && requestedCommitment.ID() == commitmentID {
-			p.networkProtocol.SendSlotCommitment(requestedCommitment, source)
-		}
-	}, event.WithWorkerPool(wpCommitments))
-
-	p.Events.Network.SlotCommitmentReceived.Hook(func(commitment *model.Commitment, source network.PeerID) {
-		p.ChainManager.ProcessCommitmentFromSource(commitment, source)
-	}, event.WithWorkerPool(wpCommitments))
-
-	p.Events.ChainManager.RequestCommitment.Hook(func(commitmentID iotago.CommitmentID) {
-		p.networkProtocol.RequestSlotCommitment(commitmentID)
-	}, event.WithWorkerPool(wpCommitments))
-
-	wpAttestations := p.Workers.CreatePool("NetworkEvents.Attestations", 1) // Using just 1 worker to avoid contention
-
-	p.Events.Network.AttestationsRequestReceived.Hook(p.ProcessAttestationsRequest, event.WithWorkerPool(wpAttestations))
-	p.Events.Network.AttestationsReceived.Hook(p.ProcessAttestations, event.WithWorkerPool(wpAttestations))
-}
-
 func (p *Protocol) initEngineManager() {
 	p.engineManager = enginemanager.New(
 		p.Workers.CreateGroup("EngineManager"),
@@ -331,160 +279,6 @@ func (p *Protocol) ProcessBlock(block *model.Block, src network.PeerID) error {
 	return nil
 }
 
-func (p *Protocol) ProcessAttestationsRequest(commitmentID iotago.CommitmentID, src network.PeerID) {
-	mainEngine := p.MainEngineInstance()
-
-	if mainEngine.Storage.Settings().LatestCommitment().Index() < commitmentID.Index() {
-		return
-	}
-
-	commitment, err := mainEngine.Storage.Commitments().Load(commitmentID.Index())
-	if err != nil {
-		return
-	}
-
-	if commitment.ID() != commitmentID {
-		return
-	}
-
-	store, err := mainEngine.Attestation.Get(commitmentID.Index())
-	if err != nil {
-		return
-	}
-
-	attestations := make([]*iotago.Attestation, 0)
-	if err := store.Stream(func(_ iotago.AccountID, attestation *iotago.Attestation) bool {
-		attestations = append(attestations, attestation)
-
-		return true
-	}); err != nil {
-		return
-	}
-
-	p.networkProtocol.SendAttestations(commitment, attestations, src)
-}
-
-func (p *Protocol) ProcessAttestations(commitment *model.Commitment, attestations []*iotago.Attestation, src network.PeerID) {
-	if len(attestations) == 0 {
-		p.Events.Error.Trigger(errors.Errorf("received attestations from peer %s are empty", src.String()))
-		return
-	}
-
-	fork, exists := p.ChainManager.ForkByForkingPoint(commitment.ID())
-	if !exists {
-		p.Events.Error.Trigger(errors.Errorf("failed to get forking point for commitment %s", commitment.ID()))
-		return
-	}
-
-	mainEngine := p.MainEngineInstance()
-
-	snapshotTargetIndex := fork.ForkingPoint.Index() - 1
-
-	snapshotTargetCommitment, err := mainEngine.Storage.Commitments().Load(snapshotTargetIndex)
-	if err != nil {
-		p.Events.Error.Trigger(errors.Wrapf(err, "failed to get commitment at snapshotTargetIndex %s", snapshotTargetIndex))
-		return
-	}
-
-	calculatedCumulativeWeight := snapshotTargetCommitment.CumulativeWeight()
-
-	visitedIdentities := make(map[iotago.AccountID]types.Empty)
-	var blockIDs iotago.BlockIDs
-	for _, att := range attestations {
-		if valid, err := att.VerifySignature(); !valid {
-			if err != nil {
-				p.Events.Error.Trigger(errors.Wrapf(err, "error validating attestation signature provided by %s", src))
-				return
-			}
-
-			p.Events.Error.Trigger(errors.Errorf("invalid attestation signature provided by %s", src))
-			return
-		}
-
-		if _, alreadyVisited := visitedIdentities[att.IssuerID]; alreadyVisited {
-			p.Events.Error.Trigger(errors.Errorf("invalid attestation from source %s, issuerID %s contains multiple attestations", src, att.IssuerID))
-			// TODO: ban source!
-			return
-		}
-
-		// TODO: this might differ if we have a Accounts with changing weights depending on the SlotIndex
-		if weight, weightExists := mainEngine.SybilProtection.Accounts().Get(att.IssuerID); weightExists {
-			calculatedCumulativeWeight += uint64(weight)
-		}
-
-		visitedIdentities[att.IssuerID] = types.Void
-
-		blockID, err := att.BlockID(mainEngine.API().SlotTimeProvider())
-		if err != nil {
-			p.Events.Error.Trigger(errors.Wrapf(err, "error calculating blockID from attestation"))
-			return
-		}
-
-		blockIDs = append(blockIDs, blockID)
-	}
-
-	// Compare the calculated cumulative weight with ours to verify it is really higher
-	weightAtForkedEventEnd := lo.PanicOnErr(mainEngine.Storage.Commitments().Load(fork.Commitment.Index())).CumulativeWeight()
-	if calculatedCumulativeWeight <= weightAtForkedEventEnd {
-		forkedEventClaimedWeight := fork.Commitment.CumulativeWeight()
-		forkedEventMainWeight := lo.PanicOnErr(mainEngine.Storage.Commitments().Load(fork.Commitment.Index())).CumulativeWeight()
-		p.Events.Error.Trigger(errors.Errorf("fork at point %d does not accumulate enough weight at slot %d calculated %d CW <= main chain %d CW. fork event detected at %d was %d CW > %d CW",
-			fork.ForkingPoint.Index(),
-			fork.Commitment.Index(),
-			calculatedCumulativeWeight,
-			weightAtForkedEventEnd,
-			fork.Commitment.Index(),
-			forkedEventClaimedWeight,
-			forkedEventMainWeight))
-		// TODO: ban source?
-		return
-	}
-
-	candidateEngine, err := p.engineManager.ForkEngineAtSlot(snapshotTargetIndex)
-	if err != nil {
-		p.Events.Error.Trigger(errors.Wrap(err, "error creating new candidate engine"))
-		return
-	}
-
-	// Set the chain to the correct forking point
-	candidateEngine.SetChainID(fork.ForkingPoint.ID())
-
-	// Attach the engine block requests to the protocol and detach as soon as we switch to that engine
-	wp := candidateEngine.Workers.CreatePool("CandidateBlockRequester", 2)
-	detachRequestBlocks := candidateEngine.Events.BlockRequester.Tick.Hook(func(blockID iotago.BlockID) {
-		p.networkProtocol.RequestBlock(blockID)
-	}, event.WithWorkerPool(wp)).Unhook
-
-	// Attach slot commitments to the chain manager and detach as soon as we switch to that engine
-	detachProcessCommitment := candidateEngine.Events.Notarization.SlotCommitted.Hook(func(details *notarization.SlotCommittedDetails) {
-		p.ChainManager.ProcessCandidateCommitment(details.Commitment)
-	}, event.WithWorkerPool(candidateEngine.Workers.CreatePool("ProcessCandidateCommitment", 2))).Unhook
-
-	p.Events.MainEngineSwitched.Hook(func(_ *engine.Engine) {
-		detachRequestBlocks()
-		detachProcessCommitment()
-	}, event.WithMaxTriggerCount(1))
-
-	// Add all the blocks from the forking point attestations to the requester since those will not be passed to the engine by the protocol
-	candidateEngine.BlockRequester.StartTickers(blockIDs)
-
-	// Set the engine as the new candidate
-	p.activeEngineMutex.Lock()
-	oldCandidateEngine := p.candidateEngine
-	p.candidateEngine = candidateEngine
-	p.activeEngineMutex.Unlock()
-
-	p.Events.CandidateEngineActivated.Trigger(candidateEngine)
-
-	if oldCandidateEngine != nil {
-		oldCandidateEngine.Shutdown()
-		if err := oldCandidateEngine.RemoveFromFilesystem(); err != nil {
-			p.Events.Error.Trigger(errors.Wrap(err, "error cleaning up replaced candidate engine from file system"))
-		}
-	}
-
-}
-
 func (p *Protocol) MainEngineInstance() *engine.Engine {
 	p.activeEngineMutex.RLock()
 	defer p.activeEngineMutex.RUnlock()
@@ -514,69 +308,5 @@ func (p *Protocol) SupportedVersions() Versions {
 func (p *Protocol) ErrorHandler() func(error) {
 	return func(err error) {
 		p.Events.Error.Trigger(err)
-	}
-}
-
-func (p *Protocol) onForkDetected(fork *chainmanager.Fork) {
-	claimedWeight := fork.Commitment.CumulativeWeight()
-	mainChainCommitment, err := p.MainEngineInstance().Storage.Commitments().Load(fork.Commitment.Index())
-	if err != nil {
-		p.Events.Error.Trigger(errors.Errorf("failed to load commitment for main chain tip at index %d", fork.Commitment.Index()))
-		return
-	}
-
-	mainChainWeight := mainChainCommitment.CumulativeWeight()
-
-	if claimedWeight <= mainChainWeight {
-		// TODO: ban source?
-		p.Events.Error.Trigger(errors.Errorf("dot not process fork with %d CW <= than main chain %d CW received from %s", claimedWeight, mainChainWeight, fork.Source))
-		return
-	}
-
-	p.networkProtocol.RequestAttestations(fork.ForkingPoint.ID(), fork.Source)
-}
-
-func (p *Protocol) switchEngines() {
-	p.activeEngineMutex.Lock()
-
-	if p.candidateEngine == nil {
-		p.activeEngineMutex.Unlock()
-		return
-	}
-
-	// Try to re-org the chain manager
-	if err := p.ChainManager.SwitchMainChain(p.candidateEngine.Storage.Settings().LatestCommitment().ID()); err != nil {
-		p.activeEngineMutex.Unlock()
-		p.Events.Error.Trigger(errors.Wrap(err, "switching main chain failed"))
-		return
-	}
-
-	if err := p.engineManager.SetActiveInstance(p.candidateEngine); err != nil {
-		p.activeEngineMutex.Unlock()
-		p.Events.Error.Trigger(errors.Wrap(err, "error switching engines"))
-		return
-	}
-
-	//// Stop current Scheduler
-	//p.CongestionControl.Shutdown()
-
-	p.linkToEngine(p.candidateEngine)
-
-	// Save a reference to the current main engine and storage so that we can shut it down and prune it after switching
-	oldEngine := p.mainEngine
-	oldEngine.Shutdown()
-
-	p.mainEngine = p.candidateEngine
-	p.candidateEngine = nil
-
-	p.activeEngineMutex.Unlock()
-
-	p.Events.MainEngineSwitched.Trigger(p.MainEngineInstance())
-
-	// TODO: copy over old slots from the old engine to the new one
-
-	// Cleanup filesystem
-	if err := oldEngine.RemoveFromFilesystem(); err != nil {
-		p.Events.Error.Trigger(errors.Wrap(err, "error removing storage directory after switching engines"))
 	}
 }
